customer_service/src/requirements.txt
pika==1.3.0
fastapi==0.79.0
pydantic==1.9.1
fastapi==0.79.0
uvicorn==0.18.3
requests==2.28.1
pymongo==4.2.0
python-dotenv==0.21.0
aiormq==6.6.4